Enable hardware watchdog  (#940)
* [C] Enable hardware watchdog for systems that have a baseboard...

+++ commandcenter/reboot_system.sh
@@ -6,9 +6,15 @@ until (( count++ >= 5 )) || ssh -o StrictHostKeyChecking=no -T $1 << EOF
 	killall pats || true
 	sleep 3
 	killall -9 pats || true
-	sudo rtcwake -m no -s 600
-	sudo swapoff -a
-	sudo systemctl poweroff
+	if [ -c /dev/baseboard ]
+	then
+		sudo rtcwake -m no -s 120
+		sudo systemctl poweroff
+	else
+		sudo rtcwake -m no -s 120
+		sudo swapoff -a
+		sudo systemctl poweroff -f
+	fi
 EOF
 do
   echo "$1 retry $count"
